pie/requirements.txt
# Stream Processing
# ------------------------------------------------------------------------------
faust-streaming==0.6.4
faust-streaming[uvloop,prometheus]==0.6.4
prometheus-client==0.11.0
kafka-python==2.0.2

# DB
# ------------------------------------------------------------------------------
asyncpg==0.23.0
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1